starts to search.
shows the search bar when you slide to the conversation list and puts everything else into the overflow menu. I choose to show it as expanded because i think it looks really cool and nice, but it would be very easy to hide it in a button as well, i already have the drawables set up, just change one boolean for that.

right now it just sets the search query to a text box, i haven't set up the layouts yet either. but i will use the query to search the database, then display the items in a list view, with every contact just mixed in there. still with your messages on the right and everyone elses on the left. I will throw their names on there like with the group mms list view.

eventually i will make it so when you click their names, it will take you to the view pager in the main activity and have it open to that persons conversation.

I am gonna need some help with this though for sure, at least querying the database and everything. it is real easy to hide this option though, just change one of the trues to false when you are on the slide out menu when you set up the options menu. So we can still put out updates while we work on getting this going, don't know how long it will take...

That was a lot tl;dr haha

I was just thinking out loud and putting some stuff in writing so you understand what i was looking for in this, and if we don't have much time the next few days, i don't forget about it lol ok thats it for this commit, work more on it tonight