added option for speed ui.
also added a new array adapter for it.

I figure this can be used with zero user customizations on the hangouts ui and it should be much faster and lag free. We just have to set them up with a default theme and go from there. I don't know if it would help to make another main activity for it or not, or if that just is unnessisary. I haven't looked at that code to see what kind of customizations it has.

Let me know if you think this is a good idea or not.