added developers tips section under settings.
This will be where we can post quick tips, just like the google+ section that i added. Users will like this and it will cut down on all the stupid emails haha

I figure i will implement it with just a quick list view and it will be very easy to add new tips to it.