converted everything to holo light theme... i don't know if this is what you want or not.
It looked much better like that on the settings menu, with the new navigation drawer, and i wanted the same look throughout the app. Seems to me like google is going more toward the holo.light.darkactionbar. so i put it in ours.

feel free to revert if this isn't what you want or you don't think it looks good. it is your app after all.