changed the soft input mode for slideover to adjustpan instead of adjust resize.
google says that this will force it so that the edit text is ALWAYS seen. Hopefully this will fix it for anyone who had custom rom problems and it wouldn't adjust correctly.

I also fixed the bug from the activity thing from John Burke. I was able to recreate his issue, then i added single top to the launch mode, and the issue is gone.