a change for notification listener... don't know if it did anything. but here are my findings:
It will not show the notification if and only if:
1.) you have the notification listener turned on
2.) you swipe sliding messaging from the recents list
3.) you DO NOT turn off your screen before receiving a message.

It seems like turning off the screen and turning it back on effectively kills the process, just like getting the broadcast does. I don't know why, don't know how, but that seems to be the case. All 3 of those things have to come into play and be exactly right, or else everything works fine. i am taking it out of the silly experimental tag that i put on it because it works just fine, for people to have all three of these happen perfectly is very unlikely. So there ya go