it works, kind of. Here are the results: For video, it doesn't seem to compress it at all, because it wasn't even able to send my 2 sec long video... i will check your code and see if you do something if forgot though
For audio, it sends completely fine and everything, but when it is received, i am not able to click on it and play it through the player... it just crashes google play music and shows a toast that it couldn't play the file.

I am able to confirm that converting them to byte arrays and everything works for the sending though, because i can play the sent messages through sliding messaging, just not the received ones.

I will look into the video thing now, and get it implemented for the regular sending, not just new compose window like it is for this commit.

You might have to take a look later though and find out why it won't receive correctly. I will test if it receives through stock correctly or not to see if that is the problem.